requirements.txt
pytube
pillow
aiohttp
aiofiles
tgcrypto
python-dotenv
py-tgcalls==0.8.1rc1
youtube-search-python
git+https://github.com/LEGEND-OS/pyrogram
git+https://github.com/LEGEND-OS/downloader
git+https://github.com/CandyMusic/yt-dl@master
